requirements.txt
TTS
soundfile
numpy
torch=2.5.1
torchvision 
torchaudio
piper-tts